+++ packages/client/src/types.tsx
@@ -1,3 +1,5 @@
+import { User } from './api/type'
+
 export type TupleUseState<T> = [T, React.Dispatch<React.SetStateAction<T>>]
 
 export type DataModalForm = Record<string, string>
@@ -22,3 +24,11 @@ export type InputAutocomlete =
   | 'login'
   | 'first_name'
   | 'second_name'
+
+export type LoadStatus = 'success' | 'loading' | 'failed' | 'noloaded'
+
+export type UserState = {
+  item: User
+  loadStatus: LoadStatus
+  isLoading: boolean
+}
